A-CodeQuality

+++ src/main/java/Store.java
@@ -2,9 +2,7 @@ import java.text.DateFormat;
 import java.text.ParseException;
 import java.text.SimpleDateFormat;
 import java.time.DateTimeException;
-import java.time.LocalTime;
 import java.io.*;
-import java.time.LocalDate;
 import java.util.ArrayList;
 import java.util.Date;
 
@@ -14,8 +12,6 @@ public class Store {
     private Integer counter;
     private ArrayList<Task> Storage;
     private DukeException DE;
-    private LocalDate LD;
-    private LocalTime LT;
     File file;
     private Ui ui = new Ui();
     private DateFormat inputDate = new SimpleDateFormat("dd/MM/yyyy");  //input date format of dd/mm/yyyy
@@ -27,7 +23,7 @@ public class Store {
      * This method updates the existing variables.
      * @param file This is the absolute file for storing the Task onto the hard disk.
      */
-    public Store(File file){
+    public Store(File file) {
         this.counter = 1;
         this.Storage = new ArrayList<>();
         this.DE = new DukeException();
@@ -39,7 +35,7 @@ public class Store {
      * @return the String of the byeMessage
      */
     public String bye() {
-        WritetoFile();
+        writeToFile();
         return ui.byeMessage();
     }
 
@@ -64,19 +60,19 @@ public class Store {
      * @param index Indicates which item on the ArrayList Storage.
      * @return String of done task.
      */
-    public String done(int index){
+    public String done(int index) {
         assert index < Storage.size() : "Index cannot be greater than list";
         assert index > 0 : "Index cannot be 0 or negative";
-//        if(index > Storage.size() || index <= 0){
-//            return DE.ExceedList();
-//        } else {
-            Task UpdateCurrAction = Storage.get(index - 1);
-            UpdateCurrAction.isDone = true;
-            UpdateCurrAction.getStatusIcon();
-            String CurrAction = UpdateCurrAction.toString();
-            WritetoFile();
+        if(index > Storage.size() || index <= 0){
+            return DE.exceedList();
+        } else {
+            Task updateCurrAction = Storage.get(index - 1);
+            updateCurrAction.isDone = true;
+            updateCurrAction.getStatusIcon();
+            String CurrAction = updateCurrAction.toString();
+            writeToFile();
             return ui.DoneMessage(CurrAction);
-//        }
+        }
     }
 
     /**
@@ -84,90 +80,62 @@ public class Store {
      * @param S The action to be done.
      * @return String This returns the number of task currently.
      */
-    public String todo(String S){
+    public String todo(String S) {
         assert S != null : "input string cannot be null";
         this.cmd = S;
-        Task T = new Todo(cmd);
-        Storage.add(T);
-        String output = T.Output() + String.format("\nNow you have %d tasks in the list.\n", Storage.size());
-        WritetoFile();
+        Task task = new Todo(cmd);
+        Storage.add(task);
+        String output = task.output()
+                        + String.format("\nNow you have %d tasks in the list.\n", Storage.size());
+        writeToFile();
         return output;
     }
 
     /**
      * This method add a new Deadline Task to the ArrayList<Task>
-     * @param ActionTime Contains the Action and the Date and/or Time.
+     * @param actionTime Contains the Action and the Date and/or Time.
      * @return the String of the new deadline task.
      */
-    public String deadline(String[] ActionTime){
-        assert ActionTime.length > 0 : "Deadline details cannot be empty";
-        String Timing;
-        String details;
-        this.cmd = ActionTime[0];
-        if(ActionTime[1].length() <=1){
-            return DE.DeadlineMissingDate();
+    public String deadline(String[] actionTime) {
+        assert actionTime.length > 0 : "Deadline details cannot be empty";
+        String timing;
+        String detail;
+        this.cmd = actionTime[0];
+        if(actionTime[1].length() <=1){
+            return DE.deadlineMissingDate();
         } else {
-            details = ActionTime[1].substring(3).strip();
-            String[] DateTime = details.split(" ");
-            try {
-                Date date = inputDate.parse(DateTime[0]);
-                Timing = outputDate.format(date);
-                if (DateTime.length == 2) {
-                    try {
-                        Date time = inputTime.parse(DateTime[1]);
-                        Timing = Timing + " " + outputTime.format(time);
-                    } catch (DateTimeException d){
-                        return DE.InvalidTimeFormat();
-                    }
-                }
-                Task T = new Deadline(cmd, Timing);
-                Storage.add(T);
-                String output = T.Output() + String.format("\nNow you have %d tasks in the list.\n", Storage.size());
-                WritetoFile();
-                return output;
-            } catch (DateTimeException | ParseException d){
-                return DE.InvalidDateFormat();
-            }
+            detail = actionTime[1].substring(3).strip();
+            timing = formatDateTime(detail);
+            Task task = new Deadline(cmd, timing);
+            Storage.add(task);
+            String output = task.output()
+                            + String.format("\nNow you have %d tasks in the list.\n", Storage.size());
+            writeToFile();
+            return output;
         }
     }
 
     /**
      * This method add a new Event Task to the ArrayList<Task>
-     * @param ActionTime Contains the Action and Date and/or Time.
+     * @param actionTime Contains the Action and Date and/or Time.
      * @return String of the event created.
      */
-    public String event(String[] ActionTime){
-        String Timing;
-        String details;
-        String Date;
-        this.cmd = ActionTime[0];
-        assert ActionTime.length > 1 : "Event missing date";
-//        if(ActionTime[1].length() <=1){
-//            return DE.EventMissingDate();
-//        } else {
-            details = ActionTime[1].substring(3).strip();
-            String[] DateTime = details.split(" ");
-            try {
-                Date date = inputDate.parse(DateTime[0]);
-                Timing = outputDate.format(date);
-                if (DateTime.length == 2) {
-                    try {
-                        Date time = inputTime.parse(DateTime[1]);
-                        Timing = date + " " + outputTime.format(time);
-
-                    } catch (DateTimeException d) {
-                        return DE.InvalidTimeFormat();
-                    }
-                }
-                Task T = new Event(cmd, Timing);
-                Storage.add(T);
-                String output = T.Output() + String.format("\nNow you have %d tasks in the list.", Storage.size());
-                WritetoFile();
-                return output;
-            } catch (DateTimeException | ParseException d) {
-                return DE.InvalidDateFormat();
-            }
-//        }
+    public String event(String[] actionTime) {
+        String timing;
+        String detail;
+        this.cmd = actionTime[0];
+        assert actionTime.length > 1 : "Event missing date";
+        if(actionTime[1].length() <=1){
+            return DE.eventMissingDate();
+        } else {
+            detail = actionTime[1].substring(3).strip();
+            timing = formatDateTime(detail);
+            Task T = new Event(cmd, timing);
+            Storage.add(T);
+            String output = T.output() + String.format("\nNow you have %d tasks in the list.", Storage.size());
+            writeToFile();
+            return output;
+        }
     }
 
     /**
@@ -175,25 +143,25 @@ public class Store {
      * @param index Indicates which item on the ArrayList<Task> Storage.
      * @return the String of the task deleted.
      */
-    public String delete(int index){
+    public String delete(int index) {
         assert index < Storage.size() : "index value greater than list";
         assert index > 0 : "index cannot be 0 or negative.";
-//        if(index > Storage.size() || index <= 0){
-//            return DE.ExceedList();
-//        } else {
-            Task UpdateCurrAction = Storage.get(index - 1);
-            String CurrAction = UpdateCurrAction.toString();
+        if(index > Storage.size() || index <= 0){
+            return DE.exceedList();
+        } else {
+            Task updateCurrAction = Storage.get(index - 1);
+            String currAction = updateCurrAction.toString();
             int amt = Storage.size() - 1;
             Storage.remove(index - 1);
-            WritetoFile();
-            return ui.DeleteMessage(CurrAction, amt);
-//        }
+            writeToFile();
+            return ui.DeleteMessage(currAction, amt);
+        }
     }
 
     /**
      * This method writes the ArrayList<Task> to the file on the hard disk.
      */
-    public void WritetoFile() {
+    public void writeToFile() {
         try {
             FileWriter fileWriter = new FileWriter(file);
             for (int i = 0; i < Storage.size(); i++) {
@@ -214,30 +182,30 @@ public class Store {
      * This method load existing file onto the ArrayList<Task> Storage.
      * @param S This is the Task from the file to be loaded from.
      */
-    public void load(String S){
+    public void load(String S) {
         boolean alreadyDone = CheckIfDone(S);
         if(S.contains("[T]")){
-            String[] result = S.split(" ", 2);
-            Task T = new Todo(result[1]);
-            T.isDone = alreadyDone;
-            Storage.add(T);
-            WritetoFile();
+            String[] results = S.split(" ", 2);
+            Task task = new Todo(results[1]);
+            task.isDone = alreadyDone;
+            Storage.add(task);
+            writeToFile();
         } else if (S.contains("[D]")){
-            String[] result = S.split(" ", 2);
-            String[] AT = result[1].split("\\|");
-            String time = AT[1].strip().substring(3).strip();
-            Task T = new Deadline(AT[0], time);
-            T.isDone = alreadyDone;
-            Storage.add(T);
-            WritetoFile();
+            String[] results = S.split(" ", 2);
+            String[] actionTime = results[1].split("\\|");
+            String time = actionTime[1].strip().substring(3).strip();
+            Task task = new Deadline(actionTime[0], time);
+            task.isDone = alreadyDone;
+            Storage.add(task);
+            writeToFile();
         } else if (S.contains("[E]")){
-            String[] result = S.split(" ", 2);
-            String[] AT = result[1].split("\\|");
-            String time = AT[1].strip().substring(3).strip();
-            Task T = new Event(AT[0], time);
-            T.isDone = alreadyDone;
-            Storage.add(T);
-            WritetoFile();
+            String[] results = S.split(" ", 2);
+            String[] actionTime = results[1].split("\\|");
+            String time = actionTime[1].strip().substring(3).strip();
+            Task task = new Event(actionTime[0], time);
+            task.isDone = alreadyDone;
+            Storage.add(task);
+            writeToFile();
         }
     }
 
@@ -246,7 +214,7 @@ public class Store {
      * @param S This is the Task.
      * @return boolean This return true if the task has been completed.
      */
-    public boolean CheckIfDone(String S){
+    public boolean CheckIfDone(String S) {
         return S.contains("[1]");
     }
 
@@ -256,25 +224,50 @@ public class Store {
      * @return the String of all the items in the list that match.
      */
     public String find(String Action){
-        ArrayList<String> Match = new ArrayList<>();
-        String output = "";
+        ArrayList<String> Matches = new ArrayList<>();
+        String output;
         int counter = 1;
         for (Task task : Storage) {
             String data = task.toString();
             if (data.contains(Action)) {
-             Match.add(String.format("%d.", counter) + data);
+             Matches.add(String.format("%d.", counter) + data);
                 counter++;
             }
         }
-        if(Match.isEmpty()){
-            output = output + DE.NoMatchesFound();
+        if(Matches.isEmpty()){
+            output = DE.noMatchesFound();
         } else {
-            output = output + "Here are the matching tasks in your list:\n";
-            for (String item : Match){
+            output = "Here are the matching tasks in your list:\n";
+            for (String item : Matches){
                 output = String.format("%s%s\n", output, item);
             }
         }
         return output;
+    }
+
+    /**
+     * This method format the input detail to date and time.
+     * @param detail This is the string to be formatted.
+     * @return the string of correct format.
+     */
+    public String formatDateTime(String detail) {
+        String timing;
+        String[] dateTime = detail.split(" ");
+        try {
+            Date date = inputDate.parse(dateTime[0]);
+            timing = outputDate.format(date);
+            if (dateTime.length == 2) {
+                try {
+                    Date time = inputTime.parse(dateTime[1]);
+                    timing = timing + " " + outputTime.format(time);
+                } catch (DateTimeException d){
+                    return DE.invalidTimeFormat();
+                }
+            }
+            return timing;
+        } catch (DateTimeException | ParseException d){
+            return DE.invalidDateFormat();
+        }
     }
 }
